feat: Like 등록/취소시 ProductId 존재여부 검사 기능 추가

+++ apps/commerce-api/src/main/java/com/loopers/application/like/LikeFacade.java
@@ -1,8 +1,10 @@
 package com.loopers.application.like;
 
-
 import com.loopers.application.like.in.LikeActionCommand;
 import com.loopers.domain.like.LikeService;
+import com.loopers.domain.product.ProductService;
+import com.loopers.support.error.CoreException;
+import com.loopers.support.error.ErrorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -10,12 +12,19 @@ import org.springframework.stereotype.Component;
 @Component
 public class LikeFacade {
     private final LikeService likeService;
+    private final ProductService productService;
 
     public boolean create(LikeActionCommand command){
+        if(productService.existsById(command.refProductId())){
+            throw new CoreException(ErrorType.NOT_FOUND, "상품 ID가 존재하지 않습니다.");
+        }
         return likeService.create(LikeActionCommand.toDomain(command));
     }
 
     public boolean delete(LikeActionCommand command) {
+        if(productService.existsById(command.refProductId())){
+            throw new CoreException(ErrorType.NOT_FOUND, "상품 ID가 존재하지 않습니다.");
+        }
         return likeService.delete(LikeActionCommand.toDomain(command));
     }
 }
